-Added “Click Tune Drag” feature -The definition of the “MIC” button has been changed such that it is now lit when the mic is enabled. -The duplex “DUP” function from the single receiver version has been added to the mRX version. -Added the ability to assign static IP addresses for up to 4 radios. -Waterfall low and high threshold settings will be remembered per receiver, per band. -Spectrum Grid Max and Min values will be remembered per receiver, per band. -Added the Waterfall AGC feature by Phil, VK6APH (from the KISS Konsole project). -Added a new “Low Noise Floor” averaging mode by Warren, NR0V. -Added a separate Gain By Band form for each of the 5 radios. -Added the MIDI tab to the CAT Control page for configuring DJ Controllers. -Removed the UCB column. -Insured that the RX Only, RX Gain, and Power controls work correctly.

+++ trunk/W5WC/PowerSDR_HPSDR_mRX/Source/Console/hardware_v2.cs
@@ -942,7 +942,7 @@ namespace PowerSDR
             }
             else
 			{
-				LatchRegister(lpt_addr, PIO_IC8, DDSRESET | DDSWRB);	// Reset the DDS chip
+				LatchRegister(lpt_addr, PIO_IC8, (byte)(DDSRESET | DDSWRB));	// Reset the DDS chip
 				LatchRegister(lpt_addr, PIO_IC8, DDSWRB);					// Leave WRB high
 			}
